Improve console reporter output
It's hard to identify test boundaries in the test report when there are
no colors in it (text file, GH actions output,...). It gets more
confusing when the test function writes to stdout.

Change the console reporter to include a clear separation between tests,
one that doesn't rely on colors.
Make the test summary more evident, to stand out in a noisy log.

Signed-off-by: Rodrigo Dias Correa <rodrigo@correas.us>