Improve error message handling
When assertion macros fail, they set a message in the test context and
abort the test immediately. When an assertion macro has multiple
error messages, it just concatenates them in the test context, creating
a cluttered, hard to read, test report.

Create a new callback to be implemented by the test reporters, called
'test_message_cb'. It will be invoked once for each message sent by the
assertion macro, allowing the reporter to decide how to display multiple
messages.

Signed-off-by: Rodrigo Dias Correa <rodrigo@correas.us>