Silence the code under test (stdout/stderr)
The console report might become hard to read when the code under test
prints to stdout or stderr.

Redirect stdout and stderr of the code under test to /dev/null, making
the console reporter output easier to read. This redirection can be
disabled with the flag '-v' when running the tests.