Disable buffering for stdout
testprefix standard output has no buffering because it doesn't write to
the stdout stream. Instead, it writes directly to a file descriptor
using 'dprintf'. Most of the time, the code under test uses stdout
(buffered). As a consequence, when the test application is executed in
verbose mode (-v), it becomes hard to read this mix of buffered and
unbuffered outputs.

Disable buffering for stdout, making both outputs unbuffered. This way,
prints from the code under test are not truncated and can be seen as
expected.