Show values in hex too when UINT macros fail
When macros that check unsigned values fail, a decimal representation of
the values is included in the report. This is enough when the values
represent counters or sizes for example. This is not so good when the
values represent are error codes or any other number that makes more
sense in hexadecimal.

Change all assertion macros that compare unsigned values to include
their hexadecimal representation along with the decimal one.